refactor(domain): turn input into an immutable value type
Merge `input_basis` into `input` as a single value type. Drop the
mutable `payment_address` cache and its upgrade-mutex; `addresses()`
now recomputes from `script_` on every call. Same philosophy as the
`header`, `script` and `output` refactors.

Net diff: -230 LOC after removing `input_basis.hpp/.cpp` and the
delegate boilerplate between the two classes.

extract_embedded_script() takes `operations()` by value (was a const
ref to the script's internal cache) so it stays correct once the
script refactor (#381) lands; no behavioural change otherwise.
