+++ src/c-api/include/kth/capi/chain/input.h
@@ -65,14 +65,6 @@ kth_size_t kth_chain_input_serialized_size(kth_input_const_t self, kth_bool_t wi
 
 // Getters
 
-/** @return Owned `kth_payment_address_mut_t`, or NULL if construction/parsing fails. Caller must release non-NULL results with `kth_wallet_payment_address_destruct`. */
-KTH_EXPORT KTH_OWNED
-kth_payment_address_mut_t kth_chain_input_address(kth_input_const_t self);
-
-/** @return Owned `kth_payment_address_list_mut_t`. Caller must release with `kth_wallet_payment_address_list_destruct`. */
-KTH_EXPORT KTH_OWNED
-kth_payment_address_list_mut_t kth_chain_input_addresses(kth_input_const_t self);
-
 /** @return Borrowed `kth_output_point_const_t` view into `self`. Do not destruct; the parent object retains ownership. Invalidated by any mutation of `self`. */
 KTH_EXPORT
 kth_output_point_const_t kth_chain_input_previous_output(kth_input_const_t self);
@@ -84,6 +76,14 @@ kth_script_const_t kth_chain_input_script(kth_input_const_t self);
 KTH_EXPORT
 uint32_t kth_chain_input_sequence(kth_input_const_t self);
 
+/** @return Owned `kth_payment_address_mut_t`, or NULL if construction/parsing fails. Caller must release non-NULL results with `kth_wallet_payment_address_destruct`. */
+KTH_EXPORT KTH_OWNED
+kth_payment_address_mut_t kth_chain_input_address(kth_input_const_t self);
+
+/** @return Owned `kth_payment_address_list_mut_t`. Caller must release with `kth_wallet_payment_address_list_destruct`. */
+KTH_EXPORT KTH_OWNED
+kth_payment_address_list_mut_t kth_chain_input_addresses(kth_input_const_t self);
+
 /** @param[out] out Must point to a null `kth_script_mut_t` slot. On success, populated with an owned handle that the caller must release via `kth_chain_script_destruct`. Untouched on error. */
 KTH_EXPORT
 kth_error_code_t kth_chain_input_extract_embedded_script(kth_input_const_t self, KTH_OUT_OWNED kth_script_mut_t* out);
@@ -93,11 +93,11 @@ kth_error_code_t kth_chain_input_extract_embedded_script(kth_input_const_t self,
 
 /** @param value Borrowed input. Copied by value into the resulting object; ownership of `value` stays with the caller. */
 KTH_EXPORT
-void kth_chain_input_set_script(kth_input_mut_t self, kth_script_const_t value);
+void kth_chain_input_set_previous_output(kth_input_mut_t self, kth_output_point_const_t value);
 
 /** @param value Borrowed input. Copied by value into the resulting object; ownership of `value` stays with the caller. */
 KTH_EXPORT
-void kth_chain_input_set_previous_output(kth_input_mut_t self, kth_output_point_const_t value);
+void kth_chain_input_set_script(kth_input_mut_t self, kth_script_const_t value);
 
 KTH_EXPORT
 void kth_chain_input_set_sequence(kth_input_mut_t self, uint32_t value);
@@ -118,10 +118,10 @@ kth_bool_t kth_chain_input_is_locked(kth_input_const_t self, kth_size_t block_he
 // Operations
 
 KTH_EXPORT
-void kth_chain_input_reset(kth_input_mut_t self);
+kth_size_t kth_chain_input_signature_operations(kth_input_const_t self, kth_bool_t bip16, kth_bool_t bip141);
 
 KTH_EXPORT
-kth_size_t kth_chain_input_signature_operations(kth_input_const_t self, kth_bool_t bip16, kth_bool_t bip141);
+void kth_chain_input_reset(kth_input_mut_t self);
 
 #ifdef __cplusplus
 } // extern "C"
