+++ src/c-api/src/chain/input.cpp
@@ -90,16 +90,6 @@ kth_size_t kth_chain_input_serialized_size(kth_input_const_t self, kth_bool_t wi
 
 // Getters
 
-kth_payment_address_mut_t kth_chain_input_address(kth_input_const_t self) {
-    KTH_PRECONDITION(self != nullptr);
-    return kth::leak_if_valid(kth::cpp_ref<cpp_t>(self).address());
-}
-
-kth_payment_address_list_mut_t kth_chain_input_addresses(kth_input_const_t self) {
-    KTH_PRECONDITION(self != nullptr);
-    return kth::leak_list<kth::domain::wallet::payment_address>(kth::cpp_ref<cpp_t>(self).addresses());
-}
-
 kth_output_point_const_t kth_chain_input_previous_output(kth_input_const_t self) {
     KTH_PRECONDITION(self != nullptr);
     return &(kth::cpp_ref<cpp_t>(self).previous_output());
@@ -115,6 +105,16 @@ uint32_t kth_chain_input_sequence(kth_input_const_t self) {
     return kth::cpp_ref<cpp_t>(self).sequence();
 }
 
+kth_payment_address_mut_t kth_chain_input_address(kth_input_const_t self) {
+    KTH_PRECONDITION(self != nullptr);
+    return kth::leak_if_valid(kth::cpp_ref<cpp_t>(self).address());
+}
+
+kth_payment_address_list_mut_t kth_chain_input_addresses(kth_input_const_t self) {
+    KTH_PRECONDITION(self != nullptr);
+    return kth::leak_list<kth::domain::wallet::payment_address>(kth::cpp_ref<cpp_t>(self).addresses());
+}
+
 kth_error_code_t kth_chain_input_extract_embedded_script(kth_input_const_t self, KTH_OUT_OWNED kth_script_mut_t* out) {
     KTH_PRECONDITION(self != nullptr);
     KTH_PRECONDITION(out != nullptr);
@@ -128,18 +128,18 @@ kth_error_code_t kth_chain_input_extract_embedded_script(kth_input_const_t self,
 
 // Setters
 
-void kth_chain_input_set_script(kth_input_mut_t self, kth_script_const_t value) {
-    KTH_PRECONDITION(self != nullptr);
-    KTH_PRECONDITION(value != nullptr);
-    auto const& value_cpp = kth::cpp_ref<kth::domain::chain::script>(value);
-    kth::cpp_ref<cpp_t>(self).set_script(value_cpp);
-}
-
 void kth_chain_input_set_previous_output(kth_input_mut_t self, kth_output_point_const_t value) {
     KTH_PRECONDITION(self != nullptr);
     KTH_PRECONDITION(value != nullptr);
     auto const& value_cpp = kth::cpp_ref<kth::domain::chain::output_point>(value);
     kth::cpp_ref<cpp_t>(self).set_previous_output(value_cpp);
+}
+
+void kth_chain_input_set_script(kth_input_mut_t self, kth_script_const_t value) {
+    KTH_PRECONDITION(self != nullptr);
+    KTH_PRECONDITION(value != nullptr);
+    auto const& value_cpp = kth::cpp_ref<kth::domain::chain::script>(value);
+    kth::cpp_ref<cpp_t>(self).set_script(value_cpp);
 }
 
 void kth_chain_input_set_sequence(kth_input_mut_t self, uint32_t value) {
@@ -169,16 +169,16 @@ kth_bool_t kth_chain_input_is_locked(kth_input_const_t self, kth_size_t block_he
 
 // Operations
 
-void kth_chain_input_reset(kth_input_mut_t self) {
-    KTH_PRECONDITION(self != nullptr);
-    kth::cpp_ref<cpp_t>(self).reset();
-}
-
 kth_size_t kth_chain_input_signature_operations(kth_input_const_t self, kth_bool_t bip16, kth_bool_t bip141) {
     KTH_PRECONDITION(self != nullptr);
     auto const bip16_cpp = kth::int_to_bool(bip16);
     auto const bip141_cpp = kth::int_to_bool(bip141);
     return kth::cpp_ref<cpp_t>(self).signature_operations(bip16_cpp, bip141_cpp);
+}
+
+void kth_chain_input_reset(kth_input_mut_t self) {
+    KTH_PRECONDITION(self != nullptr);
+    kth::cpp_ref<cpp_t>(self).reset();
 }
 
 } // extern "C"
