+++ src/domain/include/kth/domain/chain/input.hpp
@@ -11,11 +11,6 @@
 #include <memory>
 #include <vector>
 
-#if defined(__EMSCRIPTEN__)
-#include <shared_mutex>
-#endif
-
-#include <kth/domain/chain/input_basis.hpp>
 #include <kth/domain/chain/output_point.hpp>
 #include <kth/domain/chain/script.hpp>
 #include <kth/domain/define.hpp>
@@ -24,36 +19,30 @@
 #include <kth/infrastructure/math/hash.hpp>
 #include <kth/infrastructure/utility/container_sink.hpp>
 #include <kth/infrastructure/utility/reader.hpp>
-#include <kth/infrastructure/utility/thread.hpp>
 #include <kth/infrastructure/utility/writer.hpp>
-
 
 #include <kth/domain/concepts.hpp>
 
 namespace kth::domain::chain {
 
-struct KD_API input : input_basis {
-public:
+struct KD_API input {
     using list = std::vector<input>;
 
     // Constructors.
     //-------------------------------------------------------------------------
+
     input() = default;
-    using input_basis::input_basis; // inherit constructors from input_basis
+    input(output_point const& previous_output, chain::script const& script, uint32_t sequence);
+    input(output_point&& previous_output, chain::script&& script, uint32_t sequence);
 
-    explicit
-    input(input_basis const& x);
-
-    explicit
-    input(input_basis&& x) noexcept;
-
-    // Special member functions.
+    // Operators.
     //-------------------------------------------------------------------------
 
-    input(input const& x);
-    input(input&& x) noexcept;
-    input& operator=(input&& x) noexcept;
-    input& operator=(input const& x);
+    friend
+    bool operator==(input const&, input const&);
+
+    friend
+    bool operator!=(input const&, input const&);
 
     // Deserialization.
     //-------------------------------------------------------------------------
@@ -61,37 +50,86 @@ public:
     static
     expect<input> from_data(byte_reader& reader, bool wire);
 
-    // Properties (size, accessors, cache).
+    [[nodiscard]]
+    bool is_valid() const;
+
+    // Serialization.
     //-------------------------------------------------------------------------
+
+    [[nodiscard]]
+    data_chunk to_data(bool wire = true) const;
+
+    void to_data(data_sink& stream, bool wire = true) const;
+
+    template <typename W>
+    void to_data(W& sink, bool wire = true) const {
+        previous_output_.to_data(sink, wire);
+        script_.to_data(sink, true);
+        sink.write_4_bytes_little_endian(sequence_);
+    }
+
+    // Properties (size, accessors).
+    //-------------------------------------------------------------------------
+
+    /// This accounts for wire, but does not read or write it.
+    [[nodiscard]]
+    size_t serialized_size(bool wire = true) const;
+
+    output_point& previous_output();
+
+    [[nodiscard]]
+    output_point const& previous_output() const;
+
+    void set_previous_output(output_point const& value);
+    void set_previous_output(output_point&& value);
+
+    // [[deprecated]] // unsafe
+    chain::script& script();
+
+    [[nodiscard]]
+    chain::script const& script() const;
+
     void set_script(chain::script const& value);
     void set_script(chain::script&& value);
 
+    [[nodiscard]]
+    uint32_t sequence() const;
+
+    void set_sequence(uint32_t value);
+
     /// The first payment address extracted (may be invalid).
+    /// NOTE: not cached — recomputed on every call.
+    [[nodiscard]]
     wallet::payment_address address() const;
 
     /// The payment addresses extracted from this input as a standard script.
+    /// NOTE: not cached — caller owns any caching it needs.
+    [[nodiscard]]
     wallet::payment_address::list addresses() const;
 
-// protected:
+    // Validation.
+    //-------------------------------------------------------------------------
+
+    [[nodiscard]]
+    bool is_final() const;
+
+    [[nodiscard]]
+    bool is_locked(size_t block_height, uint32_t median_time_past) const;
+
+    [[nodiscard]]
+    size_t signature_operations(bool bip16, bool bip141) const;
+
+    [[nodiscard]]
+    expect<chain::script> extract_embedded_script() const;
+
     void reset();
-protected:
-    void invalidate_cache() const;
 
 private:
-    using addresses_ptr = std::shared_ptr<wallet::payment_address::list>;
-    addresses_ptr addresses_cache() const;
-
-#if ! defined(__EMSCRIPTEN__)
-    mutable upgrade_mutex mutex_;
-#else
-    mutable std::shared_mutex mutex_;
-#endif
-
-    mutable addresses_ptr addresses_;
+    output_point previous_output_;
+    chain::script script_;
+    uint32_t sequence_{0};
 };
 
 } // namespace kth::domain::chain
 
-//#include <kth/domain/concepts_undef.hpp>
-
-#endif
+#endif // KTH_DOMAIN_CHAIN_INPUT_HPP
